% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_applicable_to_coveset.R
\name{are_applicable_to_coveset}
\alias{are_applicable_to_coveset}
\alias{are_applicable_to_coveset.eqs_bag}
\title{Vectorized check equation for covariates}
\usage{
are_applicable_to_coveset(x, cov)

\method{are_applicable_to_coveset}{eqs_bag}(x, cov)
}
\arguments{
\item{x}{(eqs_bag) object}

\item{cov}{(chr) vector of covariates}
}
\value{
(lgl) for all the \code{\link{eqs}} part of \code{x},
TRUE if the covariates supplied are a subset of the
covariates needed to solve at least one of the equations in
\code{x}.
}
\description{
check which of the given \code{\link{eqs}} set of
\code{\link{eq}}uation(s) (if any) can be applied to the supplied
set of covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eqs_bag}: Check which of the
\code{\link{eqs}} set in the \code{\link{eqs_bag}} can be
applied to the \code{cov}ariates supplied.
}}

