% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_20190412}
\alias{data_20190412}
\title{data from CaloricEstimationEqu_DATA_2019-04-12_1408.csv}
\format{
An object of class \code{data.frame} with 88 rows and 128 columns.
}
\usage{
data_20190412
}
\description{
data from CaloricEstimationEqu_DATA_2019-04-12_1408.csv
}
\keyword{datasets}
