% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq}
\alias{eq}
\title{Equation}
\usage{
eq(..., name, outcome, stratum = NA_character_)
}
\arguments{
\item{...}{sequence of named coefficients}

\item{name}{(chr) the name of the equation}

\item{outcome}{(chr) the outcome produced by the equation}

\item{stratum}{(chr, default = NA) if the equation is one of a set
of equations defined each one for a level of a strata variable
\code{stratum} is the name of the reference level for this equation}
}
\value{
an \code{\link{eq}} object
}
\description{
Costructor for object of class \code{\link{eq}} defined by its
(named) coefficients.
}
\examples{
eq1 <- eq(sex = 2, age = 0.3, bmi = -0.21,
    name = "cl_test_1",
    outcome = "kcal/day"
)
}
