% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equationer-package.R
\docType{package}
\name{equationer-package}
\alias{equationer}
\alias{equationer-package}
\title{equationer: Equations of Energy Requirements}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The goal of the package `equationer` is to
    implement an environment to consistently generate the estimations
    from the equations reported in the paper "Review of Equations of
    Energy Requirements" by the *Unit of Biostatistics, Epidemiology and
    Public Health* of the University of Padova.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/equationer}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/equationer/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

}
\keyword{internal}
