% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\title{get covariates}
\usage{
get_covariates(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) covariates of the equation \code{x}
}
\description{
Extract the equation(s) covariates
}
