% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eq_names.R
\name{get_eq_names}
\alias{get_eq_names}
\alias{get_eq_names.eqs}
\title{get equations' names for a set of them}
\usage{
get_eq_names(x)

\method{get_eq_names}{eqs}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) set of names for the equations in \code{x}
}
\description{
Extract the equation(s) name(s) from bags of equations (objects of
class \code{\link{eqs}})
}
\section{Methods (by class)}{
\itemize{
\item \code{eqs}: Extract the equations's names of the given
\code{eqs} object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
get_eq_names(eqs_test)
}
