% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_update.R
\name{get_last_update}
\alias{get_last_update}
\alias{get_last_update.default}
\title{get the last update data}
\usage{
get_last_update(x)

\method{get_last_update}{default}(x)
}
\arguments{
\item{x}{(eqs_bag) the object of class (\code{\link{eqs_bag}})}
}
\value{
(date) last update date
}
\description{
Extract the last \code{\link{eqs_bag}} date
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: don't work for object different from
\code{\link{eqs_bag}}.
}}

