% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name.R
\name{get_name}
\alias{get_name}
\alias{get_name.eq}
\title{get name}
\usage{
get_name(x)

\method{get_name}{eq}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) name of \code{x}
}
\description{
Extract the equation(s) name(s)
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the name of the given
\code{eq}uation object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day"
)
get_name(eq_test)
}
