% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference.R
\name{get_reference}
\alias{get_reference}
\title{get reference}
\usage{
get_reference(x)
}
\arguments{
\item{x}{(eqs) the object of class (\code{\link{eqs}})}
}
\value{
(chr) reference of the equations' set \code{x}
}
\description{
Extract the equation(s) reference
}
