% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strata.R
\name{get_strata}
\alias{get_strata}
\alias{get_strata.eq}
\title{get strata}
\usage{
get_strata(x)

\method{get_strata}{eq}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) strata of the equation \code{x}
}
\description{
Extract the equation(s) strata
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the strata of the given
\code{eq}uation object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
get_outcome(eq_test)

eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female", nyha = 1)
)
get_outcome(eq2_test)
}
