% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R, R/get_stratum.R
\name{get_covariates.eq}
\alias{get_covariates.eq}
\alias{get_stratum}
\alias{get_stratum.eq}
\title{get stratum}
\usage{
\method{get_covariates}{eq}(x)

get_stratum(x)

\method{get_stratum}{eq}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) stratum of the equation \code{x}
}
\description{
Extract the equation(s) stratum
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the covariates of the given
\code{eq}uation object.

\item \code{eq}: Extract the stratum of the given
\code{eq}uation object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    stratum = c(sex = "female")
)
get_outcome(eq_test)
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    stratum = c(sex = "female")
)
get_outcome(eq_test)
}
