% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reer}
\alias{reer}
\title{Equations of Energy Requirements.}
\format{An object of class \code{\link{eqs_bag}}}
\source{
"Gregori et.al. 2019, Review of Equations of Energy Requirements"
}
\usage{
reer
}
\description{
All the (grouped) equations from the review made by UBESP of the
Equations of Energy Requirements.
}
\keyword{datasets}
