% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_are_applicable_to_covset.R
\name{which_are_applicable_to_covset}
\alias{which_are_applicable_to_covset}
\alias{which_are_applicable_to_covset.eqs_bag}
\title{Find equations in eqs_bag for covariates}
\usage{
which_are_applicable_to_covset(x, cov)

\method{which_are_applicable_to_covset}{eqs_bag}(x, cov)
}
\arguments{
\item{x}{(eqs_bag) object}

\item{cov}{(chr) vector of covariates}
}
\value{
(chr) names of all the \code{\link{eqs}} part of \code{x},
for which the covariates supplied are a subset of the
covariates needed to solve at least one of the equations into it.
}
\description{
Extract which of the given \code{\link{eqs}} set of
\code{\link{eq}}uation(s) (if any) can be applied to the supplied
set of covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eqs_bag}: Find which of the
\code{\link{eqs}} set in the \code{\link{eqs_bag}} can be
applied to the \code{cov}ariates supplied.
}}

