% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R, R/print_-.R
\name{eq}
\alias{eq}
\alias{print.eq}
\title{Single equation object}
\usage{
eq(..., name, outcome, strata = vector("list"))

\method{print}{eq}(x, ...)
}
\arguments{
\item{...}{sequence of named coefficients}

\item{name}{(chr) the name of the equation}

\item{outcome}{(chr) the outcome produced by the equation}

\item{strata}{(list, default = empty list) if the equation is one of
a set of equations defined each one for a compbination of
level of some strata variables, \code{strata} is a list for those
levels (named with strata names)}

\item{x}{an object used to select a method}
}
\value{
an \code{\link{eq}} object

invisible \code{x}
}
\description{
Constructor for object of class \code{\link{eq}} defined by its
(named) coefficients.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Method to print \code{\link{eq}}uations
}}

\examples{
eq1 <- eq(female = 2, age = 0.3, bmi = -0.21,
    name = "cl_test_1",
    outcome = "kcal/day"
)

# works with strata
eq2 <- eq(age = 0.3, bmi = -0.5,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq3 <- eq(age = 0.5, bmi = -0.3,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

# works with multiple strata
eq4 <- eq(age = 0.3, bmi = -0.5,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "female", mellitus = 0)
)
eq5 <- eq(age = 0.5, bmi = -0.3,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "female", mellitus = 0)
)
eq6 <- eq(age = 0.3, bmi = -0.3,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "male", mellitus = 1)
)
eq7 <- eq(age = 0.5, bmi = -0.5,
    name = "cl_test_1",
    outcome = "kcal/day",
    strata = list(sex = "male", mellitus = 1)
)

eq(age = 0.5, bmi = -0.3, name = "eq_test", outcome = "kcal/day")
}
