% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqs.R
\name{eqs}
\alias{eqs}
\title{Multiple equation object}
\usage{
eqs(..., name, reference = NA_character_)
}
\arguments{
\item{...}{(eq, only) sequence of \code{\link{eq}}uations}

\item{name}{(chr) the name of the equations' set}

\item{reference}{(chr, default NA) an optional reference for the
set of the equations in the bag.}
}
\value{
an \code{\link{eqs}} object
}
\description{
Costructor for object of class \code{\link{eqs}} defined by its
\code{\link{eq}}s components.
}
\details{
All the \code{\link{eq}}uations included in the
\code{\link{eqs}} object must share the same covariates and
the same strata (with different combinations!)
}
\examples{
eq1 <- eq(age = 0.3, bmi = -0.5,
    name    = "cl_test_1",
    outcome = "kcal/day",
    strata  = list(sex = "male")
)
eq2 <- eq(age = 0.5, bmi = -0.3,
    name = "cl_test_2",
    outcome = "kcal/day",
    strata  = list(sex = "female")
)

eqs1 <- eqs(eq1, eq2, name = "gendered-1")



# works with multiple strata
eq4 <- eq(age = 0.3, bmi = -0.5,
    name = "cl_test_4",
    outcome = "kcal/day",
    strata = list(sex = "male", mellitus = 0)
)
eq5 <- eq(age = 0.5, bmi = -0.3,
    name = "cl_test_5",
    outcome = "kcal/day",
    strata = list(sex = "female", mellitus = 0)
)
eq6 <- eq(age = 0.3, bmi = -0.3,
    name = "cl_test_6",
    outcome = "kcal/day",
    strata = list(sex = "male", mellitus = 1)
)
eq7 <- eq(age = 0.5, bmi = -0.5,
    name = "cl_test_7",
    outcome = "kcal/day",
    strata = list(sex = "female", mellitus = 1)
)

eqs2 <- eqs(eq4, eq5, eq6, eq7, name = "multistrata-1")
}
