% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_at.R
\name{evaluate_at}
\alias{evaluate_at}
\alias{evaluate_at.eq}
\title{Evaluate equation}
\usage{
evaluate_at(x, ...)

\method{evaluate_at}{eq}(x, ...)
}
\arguments{
\item{x}{equation(s set/bag) object}

\item{...}{sequence of named coefficients's values}
}
\value{
(named dbl) the value of the outcome that solve the equation when
its the covariates assume the values supplied
}
\description{
Evaluate (i.e., solve for the outcome) the supplied equation at the
supplied values of the covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Evaluate (i.e., solve for the outcome) the
supplied equation at the supplied values of the covariates.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "out"
)

evaluate_at(eq_test, age = 38, bmi = 18) # c(out = -1.6)
evaluate_at(eq_test, age = 38, sex = 1, bmi = 18) # c(out = -1.6)

\dontrun{
    # ERRORS
    evaluate_at(eq_test, age = 38, bmi = 18, sex = "female"))
    evaluate_at(eq_test, age = 38)
}
}
