% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\alias{get_covariates.eq}
\alias{get_covariates.eqs}
\alias{get_covariates.eqs_bag}
\title{get covariates}
\usage{
get_covariates(x)

\method{get_covariates}{eq}(x)

\method{get_covariates}{eqs}(x)

\method{get_covariates}{eqs_bag}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) covariates of the equation \code{x}
}
\description{
Extract the equation(s) covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the covariates of the given
\code{eq}uation object.

\item \code{eqs}: Extract the covariates of the given
\code{eq}uation object.

\item \code{eqs_bag}: Extract all the possible covariates used
by any \code{\link{eq}}uation in the \code{\link{eqs_bag}}.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
get_covariates(eq_test)
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
get_covariates(eqs_test)
}
