% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariates.R
\name{get_covariates}
\alias{get_covariates}
\alias{get_covariates.eq}
\title{get covariates}
\usage{
get_covariates(x)

\method{get_covariates}{eq}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) covariates of the equation \code{x}
}
\description{
Extract the equation(s) covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the covariates of the given
\code{eq}uation object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
get_outcome(eq_test)
}
