% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outcome.R
\name{get_outcome}
\alias{get_outcome}
\alias{get_outcome.eq}
\title{get outcome}
\usage{
get_outcome(x)

\method{get_outcome}{eq}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) outcomes of the equation \code{x} (dependent variable)
}
\description{
Extract the equation(s) outcome(s)
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Extract the outcome of the given
\code{eq}uation object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day"
)
get_outcome(eq_test)
}
