% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_update.R, R/get_reference.R,
%   R/get_strata.R
\name{get_last_update.eqs_bag}
\alias{get_last_update.eqs_bag}
\alias{get_reference.default}
\alias{get_reference.eqs}
\alias{get_reference.eqs_bag}
\alias{get_strata}
\alias{get_strata.eq}
\alias{get_strata.eqs}
\alias{get_strata.eqs_bag}
\title{get strata}
\usage{
\method{get_last_update}{eqs_bag}(x)

\method{get_reference}{default}(x)

\method{get_reference}{eqs}(x)

\method{get_reference}{eqs_bag}(x)

get_strata(x)

\method{get_strata}{eq}(x)

\method{get_strata}{eqs}(x)

\method{get_strata}{eqs_bag}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) levels' strata for the \code{\link{eq}}uation \code{x}

(list) of the strata (factors) for the
\code{\link{eq}}uations included in the \code{\link{eqs}}
\code{x}.
}
\description{
Extract the equation(s) strata
}
\section{Methods (by class)}{
\itemize{
\item \code{eqs_bag}: Extract the last date in which
\code{\link{eqs_bag}} has been updated.

\item \code{default}: Extract the strata of the given
\code{\link{eq}}uation object.

\item \code{eqs}: Extract the strata of the given
\code{\link{eqs}} object.

\item \code{eqs_bag}: Extract the strata of the given
\code{eqs_bag} object.

\item \code{eq}: Extract the strata of the given
\code{eq}uation object.

\item \code{eqs}: Extract the strata of the given
\code{eqs} object.

\item \code{eqs_bag}: Extract and merge the levels for all the
strata in any \code{\link{eq}}uations of the
\code{\link{eqs_bag}} object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
get_outcome(eq_test)

eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "female", nyha = 1)
)
get_strata(eq2_test)
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
get_strata(eqs_test)
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

eq3_test <- eq(age = -0.1, bmi = 0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "male", nyha = 1)
)
eq4_test <- eq(age = -0.1, bmi = 0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "unknown", nyha = 2)
)

eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
eqs2_test <- eqs(eq3_test, eq4_test, name = "eqs2-test")

eqs_bag1_test <- eqs_bag(eqs_test, eqs2_test,
    name = "a",
    reference = "b"
)
get_strata(eqs_bag1_test)
}
