% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference.R, R/get_strata.R
\name{get_reference.default}
\alias{get_reference.default}
\alias{get_reference.eqs}
\alias{get_strata}
\alias{get_strata.eq}
\alias{get_strata.eqs}
\title{get strata}
\usage{
\method{get_reference}{default}(x)

\method{get_reference}{eqs}(x)

get_strata(x)

\method{get_strata}{eq}(x)

\method{get_strata}{eqs}(x)
}
\arguments{
\item{x}{the object}
}
\value{
(chr) strata of the equation \code{x}
}
\description{
Extract the equation(s) strata
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Extract the strata of the given
\code{eq}uation object.

\item \code{eqs}: Extract the strata of the given
\code{eq}uation object.

\item \code{eq}: Extract the strata of the given
\code{eq}uation object.

\item \code{eqs}: Extract the strata of the given
\code{eqs} object.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
get_outcome(eq_test)

eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "female", nyha = 1)
)
get_outcome(eq2_test)
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq_test",
    outcome = "kcal/day",
    strata = list(sex = "female")
)
eq2_test <- eq(age = 0.1, bmi = -0.3,
    name    = "eq2_test",
    outcome = "kcal/day",
    strata = list(sex = "male")
)

eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
get_strata(eqs_test)
}
