% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_applicable_to_covset.R
\name{is_applicable_to_covset}
\alias{is_applicable_to_covset}
\alias{is_applicable_to_covset.eq}
\title{Check equation for covariates}
\usage{
is_applicable_to_covset(x, cov)

\method{is_applicable_to_covset}{eq}(x, cov)
}
\arguments{
\item{x}{equation(s set/bag) object}

\item{cov}{(chr) vector of covariates}
}
\value{
(lgl) TRUE if the covariates supplied are a subset of the
covariates needed to solve at least one of the equations in
\code{x}.
}
\description{
Check if any of the given equation(s) can be applied to the supplied
set of covariates
}
\section{Methods (by class)}{
\itemize{
\item \code{eq}: Check if the \code{eq}uation can
be applied to \code{cov}.
}}

\examples{
library(equationer)
eq_test <- eq(age = 0.1, bmi = -0.3,
    name    = "first eq_test",
    outcome = "kcal/day"
)
is_applicable_to_covset(eq_test, c("age", "bmi")) # TRUE
is_applicable_to_covset(eq_test, c("age", "bmi", "sex")) # TRUE
is_applicable_to_covset(eq_test, c("age")) # FALSE
is_applicable_to_covset(eq_test, c("age", "sex")) # FALSE
}
